Refactor namespaces and add dictionary extensions
Updated namespaces in multiple files from `Startup.Common.Helpers.Extensions` to `Startup.Common.Helpers.Data` to improve code organization. Refactored `LoggerSupport.cs` to align with the new namespace structure. Added `DictionaryExtensions.cs` with `GetOrAdd` and `TryUpdate` methods to enhance dictionary operations.
